Tests: Move all multiprocessing tests to one file.
